Add more readable table output
