+++ pkg/testrunner/result/collect.go
@@ -16,13 +16,13 @@ package result
 
 import (
 	"fmt"
+	"path/filepath"
+
 	"github.com/gardener/gardener/pkg/client/kubernetes"
 	tmv1beta1 "github.com/gardener/test-infra/pkg/apis/testmachinery/v1beta1"
 	"github.com/gardener/test-infra/pkg/testrunner"
 	"github.com/gardener/test-infra/pkg/util"
 	log "github.com/sirupsen/logrus"
-	"path/filepath"
-	"sort"
 )
 
 // Collect collects results of all testruns and writes them to a file.
@@ -56,20 +56,8 @@ func Collect(config *Config, tmClient kubernetes.Interface, namespace string, ru
 			log.Errorf("Testrun %s failed with phase %s", run.Testrun.Name, run.Testrun.Status.Phase)
 		}
 		fmt.Print(util.PrettyPrintStruct(run.Testrun.Status))
+		printStatusTable(run.Testrun.Status.Steps)
 	}
 
 	return testrunsFailed, nil
 }
-
-// orderSteps orders the steps by their finihsed date
-func orderSteps(steps []*tmv1beta1.StepStatus) {
-	sort.Sort(StepStatusList(steps))
-}
-
-type StepStatusList []*tmv1beta1.StepStatus
-
-func (s StepStatusList) Less(a, b int) bool {
-	return s[a].StartTime.Before(s[b].StartTime)
-}
-func (s StepStatusList) Len() int      { return len(s) }
-func (s StepStatusList) Swap(a, b int) { s[a], s[b] = s[b], s[a] }
